+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Please make sure to add your changes to the appropriate categories:
 - Made `ProgressId` impl `Default`.
 - Made `Report` impl `Default`.
 - Added `to_pruned()`/`into_pruned()` methods to `Report` for obtaining a pruned report containing only those sub-reports that were updated since a certain generation and thus need updating of their corresponding user-facing UI.
+- Added `as_raw()` method to `Generation` for accessing the raw internal generational counter value.
 
 ### Changed